requirements.txt
pandas>=2.2.2
python-docx>=1.1.2
reportlab>=4.2.2
PyYAML>=6.0.2
openpyxl>=3.1.5